+++ src/components/App/Handle.jsx
@@ -12,6 +12,7 @@ import { showAlert } from '@/reducers/app/action';
 import { signUp, fetchProfile, authUpdate, getFreeETH } from '@/reducers/auth/action';
 import { getUserProfile } from '@/reducers/exchange/action';
 import { createMasterWallets } from '@/reducers/wallet/action';
+import { getListOfferPrice } from '@/reducers/exchange/action';
 // components
 import { MasterWallet } from '@/models/MasterWallet';
 import Loading from '@/components/core/presentation/Loading';
@@ -32,6 +33,7 @@ class Handle extends React.Component {
     //
     getFreeETH: PropTypes.func.isRequired,
     refer: PropTypes.string,
+    getListOfferPrice: PropTypes.func.isRequired,
   }
 
   static defaultProps = {
@@ -62,6 +64,16 @@ class Handle extends React.Component {
     }
     return null;
   }
+
+  // exchange
+  getListOfferPrice() {
+    const ipInfo = local.get(APP.IP_INFO);
+    this.props.getListOfferPrice({
+      PATH_URL: API_URL.EXCHANGE.GET_LIST_OFFER_PRICE,
+      qs: { fiat_currency: ipInfo?.currency },
+    });
+  }
+  // /exchange
 
   // wallet
   getFreeETH() {
@@ -155,6 +167,10 @@ class Handle extends React.Component {
         // exchange
         this.props.getUserProfile({ PATH_URL: API_URL.EXCHANGE.GET_USER_PROFILE });
 
+        // exchange
+        this.getListOfferPrice();
+        this.timeOutGetPrice = setInterval(() => this.getListOfferPrice(), 2 * 60 * 1000); // 2'
+
         // wallet
         const listWallet = MasterWallet.getMasterWallet();
         // console.log('app - handle - wallet - listWallet - ', listWallet);
@@ -241,4 +257,5 @@ export default compose(withFirebase, connect(state => ({
   authUpdate,
   getUserProfile,
   getFreeETH,
+  getListOfferPrice,
 }))(Handle);
